Divide some assertions on server and client side
