+++ testing/tests/DevExpress.ui.widgets.editors/switch.markup.tests.js
@@ -1,6 +1,7 @@
 "use strict";
 
 var $ = require("jquery");
+var windowUtils = require("core/utils/window");
 
 require("common.css!");
 require("ui/switch");
@@ -17,11 +18,23 @@ QUnit.testStart(function() {
 
 var SWITCH_CLASS = "dx-switch",
     WRAPPER_CLASS = "dx-switch-wrapper",
-    CONTAINER_CLASS = "dx-switch-container";
+    CONTAINER_CLASS = "dx-switch-container",
+
+    INNER_CLASS = "dx-switch-inner",
+    INNER_SELECTOR = "." + INNER_CLASS,
+    HANDLE_CLASS = "dx-switch-handle",
+    HANDLE_SELECTOR = "." + HANDLE_CLASS,
+
+    LABEL_ON_CLASS = "dx-switch-on",
+    LABEL_OFF_CLASS = "dx-switch-off",
+    LABEL_ON_SELECTOR = "." + LABEL_ON_CLASS,
+    LABEL_OFF_SELECTOR = "." + LABEL_OFF_CLASS;
 
 QUnit.module("Switch markup");
 
 QUnit.test("markup", function(assert) {
+    var check = windowUtils.hasWindow() ? assert.ok.bind(assert) : assert.notOk.bind(assert);
+
     var element = $("#switch").dxSwitch();
 
     assert.ok(element.hasClass(SWITCH_CLASS));
@@ -32,6 +45,18 @@ QUnit.test("markup", function(assert) {
     var container = wrapper.children();
     assert.equal(container.length, 1);
     assert.ok(container.hasClass(CONTAINER_CLASS));
+
+    var inner = element.find(INNER_SELECTOR);
+    check(inner.length, "Switch inner");
+
+    var labelOnEl = element.find(LABEL_ON_SELECTOR);
+    check(labelOnEl.length, "Switch label");
+
+    var handleEl = element.find(HANDLE_SELECTOR);
+    check(handleEl.length, "Switch handle");
+
+    var labelOffEl = element.find(LABEL_OFF_SELECTOR);
+    check(labelOffEl.length, "Switch label");
 });
 
 QUnit.test("a hidden input should be rendered", function(assert) {
